cmd/luart: build against 5.5 luart, with lua.c 5.5's options
The pin moves to main at 8f5ae77, the library's 5.5 standard library and
API. As 5.5's lua.c: -W turns warnings on; -l g=mod requires mod into
global g, and -l mod-v1 into mod; LUA_INIT_5_5 comes before LUA_INIT;
the state closes at exit, running pending finalizers and flushing
files. The version line names 5.5's copyright years, and global
declarations highlight as keywords, which chroma's Lua lexer predates.

+++ cmd/luart/eval.go
@@ -151,7 +151,7 @@ func (f *formatter) table(index, depth int, seen map[any]bool) *pnode {
 	}
 	l.PushNil()
 	for l.Next(index) {
-		if k, ok := l.ToInteger(-2); ok && l.TypeOf(-2) == lua.TypeNumber && 1 <= k && k <= length {
+		if k, ok := l.ToInteger(-2); ok && l.TypeOf(-2) == lua.TypeNumber && 1 <= k && k <= int64(length) {
 			l.Pop(1)
 			continue
 		}
